[Feat] #8 github 로그인 파일 수정 - github 로그인 파일 수정 - github 로그인 파일의 경로 수정, 내부 로직 리팩토링
